Move validate_time_metrics_outside_transaction to validators directory
